Fixed all the bugs of porting to gsl_matrix usage

+++ mds.h
@@ -11,7 +11,7 @@
 
 #define SQR(x) (x)*(x)
 
-float loss_function_simple (gsl_matrix *p, gsl_matrix *d, float lim);
+double loss_function_simple (gsl_matrix *p, gsl_matrix *d, float lim);
 void step_function_internal (gsl_matrix *p, float var, gsl_rng * number_generator); /* var stands for width of distribution */
 void print_matrix_2d (gsl_matrix *p, char* str);
 #endif
